backend/requirements.txt
fastapi==0.103.1
uvicorn==0.25.0
python-dotenv==1.0.0
python-multipart==0.0.6
pydantic-settings==2.0.3
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
SQLAlchemy==2.0.20